index.ts splited, server.ts created, server is running :)
