+++ server/index.ts
@@ -1,19 +1,9 @@
 import dotenv from "dotenv";
 dotenv.config();
-import express from "express";
-import { Application } from "express";
-import cookieParser from "cookie-parser";
-import cors from "cors";
-import path from "path";
-const app: Application = express();
-const port = process.env.PORT;
 
-const router = require("./router");
-app.use(cors({ credentials: true, origin: "http://localhost:3000" }));
-app.use(express.json());
-app.use(cookieParser());
-app.use("/images", express.static(path.join(__dirname, "/images")));
-app.use(router);
+import { app } from "./server";
+
+const port = process.env.PORT;
 
 app.listen(port, () => {
   console.log(`app listening on port ${port}`);
